# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Espaço Sinapse — API administrativa",
    "description": "Sessão HttpOnly. Obtenha GET /api/v1/auth/csrf e envie seu token no headerName retornado em operações de escrita. Após login, obtenha novo CSRF. ADMIN administra cadastros; RECEPCAO gerencia pacientes e agenda. Datas usam offset ISO-8601. Atualizações exigem version. Erros usam ProblemDetail; conflitos retornam 409, incluindo conflictingAppointmentIds quando aplicável. API pública expõe somente campos aprovados.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8081",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/v1/professionals/{id}": {
      "get": {
        "tags": [
          "professional-controller"
        ],
        "operationId": "get_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "professional-controller"
        ],
        "operationId": "update_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfessionalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "professional-controller"
        ],
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/professionals/{id}/availability": {
      "get": {
        "tags": [
          "professional-controller"
        ],
        "operationId": "availability",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityInput"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "professional-controller"
        ],
        "operationId": "availability_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityInput"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/professionals": {
      "get": {
        "tags": [
          "professional-controller"
        ],
        "operationId": "list_2",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageDtoProfessionalDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "professional-controller"
        ],
        "operationId": "create_2",
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfessionalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/professionals/{id}/blocks": {
      "get": {
        "tags": [
          "professional-controller"
        ],
        "operationId": "blocks",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlockDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "professional-controller"
        ],
        "operationId": "block",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BlockDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/professionals/{id}/active": {
      "patch": {
        "tags": [
          "professional-controller"
        ],
        "operationId": "active_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActiveInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessionalDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/professionals/{id}/blocks/{blockId}": {
      "delete": {
        "tags": [
          "professional-controller"
        ],
        "operationId": "deleteBlock",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "Token retornado por GET /api/v1/auth/csrf. Conserve o cookie da mesma sessão; obtenha um novo token após login.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProfessionalInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 0
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254,
            "minLength": 0
          },
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 0
          },
          "registration": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "region": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "bio": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 0
          },
          "serviceIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "maxItems": 100,
            "minItems": 0,
            "uniqueItems": true
          },
          "published": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "name",
          "serviceIds"
        ]
      },
      "ProfessionalDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "registration": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "serviceIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "uniqueItems": true
          },
          "active": {
            "type": "boolean"
          },
          "published": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AvailabilityInput": {
        "type": "object",
        "properties": {
          "periods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PeriodInput"
            },
            "maxItems": 35,
            "minItems": 0
          }
        },
        "required": [
          "periods"
        ]
      },
      "PeriodInput": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "integer",
            "format": "int32",
            "maximum": 7,
            "minimum": 1
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          }
        },
        "required": [
          "endTime",
          "startTime"
        ]
      },
...
      "BlockInput": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "end",
          "start"
        ]
      },
      "BlockDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ActiveInput": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "active",
          "version"
        ]
      },
...
      "PageDtoProfessionalDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfessionalDto"
            }
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "name": "JSESSIONID",
        "in": "cookie"
      }
    }
  }
}